refactored prepare files back into one and some refactoring on utilties
