+++ Utilities/adjust_data.py
@@ -22,14 +22,16 @@ def get_batch_ids(path: Union[str, list[str]]) -> set:
 
 
 def remove_batch_ids(data: dict, batch_ids: set, inplace: bool = True, modify_batch_ids: bool = True) -> dict:
+    if len(batch_ids) == 0:
+        return data
     if not modify_batch_ids:
         batch_ids = batch_ids.copy()
     # create list of indices which have to be removed
     indices_to_remove = []
-    for i, batch_id in enumerate(data['batch_ids']):
-        if batch_id[0] in batch_ids:
+    for i, (batch_id,) in enumerate(data['batch_ids']):
+        if batch_id in batch_ids:
             indices_to_remove.append(i)
-            batch_ids.remove(batch_id[0])
+            batch_ids.remove(batch_id)
     # reverse list of indices to enable easy removing of items of a list by consecutive pops
     indices_to_remove.reverse()
     if inplace:
@@ -41,12 +43,12 @@ def remove_batch_ids(data: dict, batch_ids: set, inplace: bool = True, modify_ba
     for i in indices_to_remove:
         ids.pop(i)
     new_data['batch_ids'] = ids
-    new_data['batch_comp'] = np.delete(data['batch_comp'], indices_to_remove)
+    new_data['batch_comp'] = np.delete(data['batch_comp'], indices_to_remove, axis=0)
     if not inplace:
         new_data['target'] = {}
     for target in data['target']:
-        new_data['target'][target] = np.delete(data['target'][target], indices_to_remove)
-    new_data['comps'] = np.delete(data['comps'], indices_to_remove)
+        new_data['target'][target] = np.delete(data['target'][target], indices_to_remove, axis=0)
+    new_data['comps'] = np.delete(data['comps'], indices_to_remove, axis=0)
 
     return new_data
