+++ setup.cfg
@@ -13,5 +13,5 @@ package_dir = =.
 [options.entry_points]
 console_scripts =
     train-CGAT = CGAT.train:run
-    prepare = CGAT.prepare_volume_data:main
+    prepare = CGAT.prepare_data:main
     train-GP = CGAT.gaussian_process:main
